requirements.txt
# FastAPI 기반 웹 프레임워크
fastapi==0.115.0

# ASGI 서버 (개발 시 --reload, 운영 시 gunicorn+uvicorn.workers.UvicornWorker 등 조합 가능)
uvicorn[standard]==0.30.6
python-multipart

# Prometheus 메트릭 노출
prometheus-client==0.20.0

# 데이터 모델/검증 (Pydantic v2)
pydantic==2.9.2

# 이미지 EXIF 읽기
Pillow==10.4.0

# .env 파일을 환경변수로 로드 (개발 편의)
python-dotenv==1.0.1
